+++ src/DependencyInjection/ConfigFactory.php
@@ -53,7 +53,7 @@ final class ConfigFactory
         $keyFile = \is_array($config['key_file'] ?? null) ? $config['key_file'] : [];
         $serve = $config['serve_key_file'] ?? null;
         $config['serve_key_file'] = \is_bool($serve) ? $serve : (bool) ($keyFile['enabled'] ?? true);
-        unset($config['messenger'], $config['doctrine'], $config['key_file'], $config['sitemap'], $config['profiler']);
+        unset($config['messenger'], $config['doctrine'], $config['key_file'], $config['sitemap'], $config['profiler'], $config['flush']);
         if (\is_array($config['logging'] ?? null)) {
             unset($config['logging']['channel']);
         }